+++ Kerberos/src/Platform/OpenGL/OpenGLTexture.h
@@ -15,7 +15,7 @@ namespace Kerberos
 		uint32_t GetWidth() const override { return m_Width; }
 		uint32_t GetHeight() const override { return m_Height; }
 
-		uint32_t GetRendererID() const override { return m_RendererID; }
+		uint64_t GetRendererID() const override { return m_RendererID; }
 		
 		void Bind(uint32_t slot = 0) const override;
